Fix tests to stub the with https
I'm not sure what the difference is but my hypothesis is that the old
set of stubs don't work with old docker versions (I'm on 1.6.1
currently).  Could be old docker parses urls in a more naive way and
forces the protocol from tcp => https.  I was getting erros like:

  "this is stubbed"
  stub_request(:get, "http://tcp//example.com:5422/v1 ...

  "you need to stub this"
  stub_request(:get, "https://example.com:5422/v1 ...

And note this is how the docker url is set in the helper:

  Docker.url = 'tcp://example.com:5422'